requirements.txt
customtkinter>=5.2.0
sounddevice>=0.4.6
websockets>=12.0
numpy>=1.24.0
pyyaml>=6.0
pynput>=1.7.6
webrtcvad>=2.0.10
